ds-ts | Adding Queue data structure implementation
This commit consists of the Queue data structure implementation. The following methods have been added for Queue data strucure.
1. enqueue
2. dequeue
3. peek
In addition getters are provided for getting the size, capacity and whether the queue isEmpty or isFull. A capacity error is thrown if the capacity is provided and the queue exceeds the limit.