ds-ts | Updating insertNode method implementation for binary tree
This commit updates the insertNode method of the binary tree implementation to use a generator function in order to write an essentially iterative algorithm in a recursive style. This update makes our insertNode method consistent with the traversal methods updated in a previous commit to use generators to write recursive looking iterative code for improving performance.