ds-ts | Publicly exposing the root as a read-only value
This commit contains the updates allowing the root of the tree to be accessed as a read-only value. Despite the fact that an isEmpty property is already provided which would indicate whether the tree is empty or not, the root value can sometimes be useful for the consumers of the API. This descision may be revisited in the future and the root may go back to being a protected/private field inaccessible to the consumers.