ds-ts | Binary tree equality checks update
This commit contains the updates to the equality checks in the binary tree implementation. All equality checks accross the file are being done using the equals function to ensure consistency and correctness in comparison logic. This may be updated in future to use the compare function instead or use the compare function alongside the equals function in case the user does not provide a separate equals function for comparing complex data types like objects or arrays.