ds-ts | Update delete method for tree
The delete method API has been updated to return a boolean instead of void. This would keep the tree API more consistent with other standard data-structure API's and would also inform the consumer about the success or failure of the call. The delete call would return false in case there is no root, which means that the tree does not exist, and in any other case, the delete method would succeed by deleting the tree and returning true.